requirements.txt
Flask==2.2.2
requests==2.28.1
Werkzeug==2.2.2
debugpy # Required for debugging.
google-cloud-aiplatform
google-cloud
opencv-python
pytz
numpy
google-cloud-storage
pillow
scikit-image
torch
ultralytics